Separate ALTO functionality into altorenderer.cpp
