+++ CMakeLists.txt
@@ -215,6 +215,7 @@ set(tesseract_src ${tesseract_src}
     src/api/capi.cpp
     src/api/renderer.cpp
     src/api/pdfrenderer.cpp
+    src/api/altorenderer.cpp
 )
 
 if (WIN32)
@@ -223,7 +224,7 @@ if (WIN32)
         set(tesseract_hdr
             ${tesseract_hdr}
             ${CMAKE_CURRENT_SOURCE_DIR}/src/vs2010/tesseract/resource.h)
-        set(tesseract_rsc ${CMAKE_CURRENT_BINARY_DIR}/vs2010/tesseract/libtesseract.rc)
+        set(tesseract_rsc ${CMAKE_CURRENT_BINARY_DIR}/vs2010/tesseract/libtesseract.rc src/api/altorenderer.cpp)
         set_source_files_properties(
             ${CMAKE_CURRENT_SOURCE_DIR}/src/arch/dotproductsse.cpp
             PROPERTIES COMPILE_DEFINITIONS __SSE4_1__)
