+++ src/api/renderer.cpp
@@ -194,54 +194,6 @@ namespace tesseract {
     }
 
 /**********************************************************************
- * Alto Text Renderer interface implementation
- **********************************************************************/
-    TessAltoRenderer::TessAltoRenderer(const char *outputbase)
-            : TessResultRenderer(outputbase, "alto") {
-    }
-
-    bool TessAltoRenderer::BeginDocumentHandler() {
-      AppendString(
-              "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
-              "<alto xmlns=\"http://www.loc.gov/standards/alto/ns-v3#\" xmlns:xlink=\"http://www.w3.org/1999/xlink\" xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xsi:schemaLocation=\"http://www.loc.gov/standards/alto/ns-v3# http://www.loc.gov/alto/v3/alto-3-0.xsd\">\n"
-              "\t<Description>\n"
-              "\t\t<MeasurementUnit>pixel</MeasurementUnit>\n"
-              "\t\t<sourceImageInformation>\n"
-              "\t\t\t<fileName>");
-
-      AppendString(title());
-
-      AppendString("\t\t\t</fileName>\n"
-                   "\t\t</sourceImageInformation>\n"
-                   "\t\t<OCRProcessing ID=\"OCR_0\">\n"
-                   "\t\t\t<ocrProcessingStep>\n"
-                   "\t\t\t\t<processingSoftware>\n"
-                   "\t\t\t\t\t<softwareName>tesseract 4.0.0</softwareName>\n"
-                   "\t\t\t\t</processingSoftware>\n"
-                   "\t\t\t</ocrProcessingStep>\n"
-                   "\t\t</OCRProcessing>\n"
-                   "\t</Description>\n"
-                   "\t<Layout>\n");
-
-      return true;
-    }
-
-    bool TessAltoRenderer::EndDocumentHandler() {
-      AppendString("\t</Layout>\n</alto>\n");
-
-      return true;
-    }
-
-    bool TessAltoRenderer::AddImageHandler(TessBaseAPI* api) {
-      const std::unique_ptr<const char[]> hocr(api->GetAltoText(imagenum()));
-      if (hocr == nullptr) return false;
-
-      AppendString(hocr.get());
-
-      return true;
-    }
-
-/**********************************************************************
  * TSV Text Renderer interface implementation
  **********************************************************************/
     TessTsvRenderer::TessTsvRenderer(const char* outputbase)
